+++ modules/alloy/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Upgraded to Katamari 9.0, which includes breaking changes to the `Optional` API used in this module.
 
+### Fixed
+- Slider components did not respond to `Representing.setValue` #TINY-8304
+
 ## 9.0.2 - 2021-10-11
 
 ### Fixed